# perf: ingestion tweaking for bigquery