--- a/lib/logflare/backends/ingest_event_queue/queue_janitor.ex
+++ b/lib/logflare/backends/ingest_event_queue/queue_janitor.ex
@@ -15,7 +15,7 @@ defmodule Logflare.Backends.IngestEventQueue.QueueJanitor do
   @default_interval 1_000
   @default_remainder 100
   @default_max Logflare.Backends.max_buffer_queue_len()
-  @default_purge_ratio 0.1
+  @default_purge_ratio 0.05
 
   def start_link(opts) do
     GenServer.start_link(__MODULE__, opts)
@@ -70,7 +70,10 @@ defmodule Logflare.Backends.IngestEventQueue.QueueJanitor do
 
         Logger.warning(
           "IngestEventQueue private :ets buffer exceeded max for source id=#{state.source_id}, dropping #{to_drop} events",
-          backend_id: state.backend_id
+          backend_id: state.backend_id,
+          source_id: state.source_token,
+          source_token: state.source_token,
+          ingest_drop_count: to_drop
         )
       end
     end